spilt class and script to two file.
